backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
mysql-connector-python==8.2.0
python-dotenv==1.0.0
emergentintegrations==0.1.0
PyJWT==2.8.0
bcrypt==4.0.1
requests==2.31.0